fix: throw a better error when a homepage URL cannot be parsed
