+++ package.json
@@ -44,6 +44,8 @@
     "yargs": "^14.0.0"
   },
   "devDependencies": {
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
     "codecov": "^3.6.1",
     "eslint": "^6.2.1",
     "eslint-config-standard": "^14.0.1",
