+++ test/fixtures/invalid-homepage/resources/app/index.js
@@ -1,0 +1,1 @@
+console.log('BAR');
