+++ test/fixtures/invalid-homepage/version
@@ -1,0 +1,1 @@
+v0.29.2
